Enter OOP
Introducing a type to keep the state of the parser allows the
abstracting of the handle_event function and reusing it in places, where
it parse.next() is used, but the result is discarded if it's not what's
expected.

This is an improvement, although it does not fix the text. See next
commit for fix.